trx to lib
