+++ src/routes/cylinder-set/deleteCylinderSet.ts
@@ -1,12 +1,11 @@
 import { FastifyInstance, FastifyReply, FastifyRequest } from 'fastify';
-import { knexController } from '../../database/database';
 import { errorHandler } from '../../lib/errorHandler';
 import { Type } from '@sinclair/typebox';
 import {
-  Cylinder,
   cylinderSetIdParamsPayload,
   CylinderSetIdParamsPayload,
 } from '../../types/cylinderSet.types';
+import { deleteCylinderSet } from '../../lib/deleteCylinderSet';
 
 const deleteSetReply = Type.Object({
   setID: Type.String(),
@@ -37,33 +36,14 @@ const handler = async (
 
   const setID = request.params.cylinderSetId;
 
-  await knexController.transaction(async (trx) => {
-    // Get id's of cylinders that belong to given set.
-    const rowData = await trx
-      .select('cylinder')
-      .from<Cylinder>('diving_cylinder_to_set')
-      .where('cylinder_set', setID);
+  // Delete culinder set.
+  const res = await deleteCylinderSet(setID);
 
-    if (rowData.length === 0) {
-      return errorHandler(reply, 404, 'No set with given id found.');
-    }
+  if (res.status !== 200) {
+    return errorHandler(reply, res.status, res.message);
+  }
 
-    // Delete cylinders from set.
-    await trx('diving_cylinder_to_set').where('cylinder_set', setID).del();
-
-    // Delete single cylinders.
-    // Knex queries return RowDataPacket obj's, and can't get them to work with whereIn :)
-    const cylinders: string[] = Object.values(
-      JSON.parse(JSON.stringify(rowData))
-    );
-
-    await trx.from('diving_cylinder').whereIn('id', cylinders).del();
-
-    // Delete set.
-    await trx('diving_cylinder_set').where('id', setID).del();
-  });
-
-  await reply.code(200).send({ setID, message: 'Set deleted successfully!' });
+  await reply.code(res.status).send({ setID, message: res.message });
 };
 
 export default async (fastify: FastifyInstance): Promise<void> => {
